refactor: inputView, Option, 파일 처리 클래스 리팩토링
