+++ src/test/java/FileLoaderTest.java
@@ -1,12 +1,12 @@
 import static org.assertj.core.api.Assertions.assertThat;
 
 import domain.AttendanceBook;
-import domain.FileWithAttendanceData;
+import loader.FileLoader;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-public class FileWithAttendanceDataTest {
+public class FileLoaderTest {
 
     @DisplayName("12월 출석 기록 파일(attendances.csv)을 읽어 객체를 생성할 수 있다.")
     @ParameterizedTest
@@ -14,7 +14,7 @@ public class FileWithAttendanceDataTest {
     void should_CreateObject_When_givenData(String name) {
         String filePath = "src/main/resources/attendances.csv";
         AttendanceBook attendanceBook = new AttendanceBook();
-        FileWithAttendanceData fileReader = new FileWithAttendanceData(attendanceBook);
+        FileLoader fileReader = new FileLoader(attendanceBook);
 
         fileReader.loadFile(filePath);
